move api surrounding modules from webserver to bumps main package level
