+++ bumps/fit_thread.py
