+++ bumps/persistent_settings.py
