+++ bumps/plots/colors.py
