+++ bumps/plots/convergence_plot.py
