+++ bumps/plots/traceplot.py
