+++ bumps/plots/varplot.py
